More decompositions for jvp from functorch into core
[ghstack-poisoned]
